+++ RELEASE_NOTES.md
@@ -81,3 +81,8 @@
 ## 2.6.0
 
 * fix raw api requests
+
+## 2.7.0
+
+* corresponds to hafas-client version 6.2.0
+* rename field StopOver additionalStop to additional